hace_django_project/requirement.txt
django
pandas
openai
matplotlib
plotly
ipython
num2words
numpy
scipy
sklearn
scikit-learn
djangorestframework
django-cors-headers